pyproject.toml
[project]
name = "rsjbuild"
version = "0.1.13"
description = "Buuild tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "addict>=2.4.0",
    "babel>=2.17.0",
    "cryptography>=45.0.3",
    "cython>=3.1.1",
    "fabric>=3.2.2",
    "jinja2>=3.1.6",
    "mdx-truly-sane-lists>=1.3",
    "mkdocs>=1.6.1",
    "polib>=1.2.0",
    "pyjwt>=2.10.1",
    "setuptools>=80.9.0",
    "six>=1.17.0",
]

[project.scripts]
rsjbuild = "rsjbuild.rsjbuild:main"

[tool.uv]
package = true

[tool.ruff]
builtins = ["_"]
lint.ignore = ["E402"]

[tool.setuptools.package-data]
rsjbuild = ["*.pyx", "*.h"]
